requirements.txt
python-telegram-bot==21.10
yt-dlp==2025.2.19
spotipy==2.25.0
moviepy==1.0.3
python-dotenv==1.0.1